debugger: ensure processor is happy before invoke()ing agents
Targets with garbage at 0 may be in an exceptions-pending
state after reset-stop, so clear exceptions and make sure
PSR is sane, in addition to PC, SP, and friends.

A more sure way to avoid this would be to hold the target's
SYSRESETn asserted while setting up, but that's target-
specific.

See this ARM KB article:
http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.faqs/ka16214.html